Add files for lambda build and deploy
I've added in a job to build and deploy lambdas. This is specifically
for the scala ones but the antivirus deploy process is complicated and
won't fit in here anyway so making it more generic won't help.

I've also moved the get-next version script into its own action as this
is used more than once.