Add script to check deployed environment variables
This runs a python script inline:

It runs a regex search looking for the {ENV_VAR} lines in the conf file.
Then it gets the env vars from the lambda
It then compares them and fails if any are missing in the lambda.

The python is all sqashed together because I can't have it in a separate
file inside a custom action.