Update the ignore list
The AWS libraries are on 2.19.x now so we're getting updates every time as they're release every day.
I've also added the jaxb library. This is only used by the new file format as it's needed by the Droid library. This will break if upgraded so we need to stop trying.