Fix the inconsistent case.
The variable being set on the GITHUB_OUTPUT file was `title-environment`
but the outputs which were set to the next stage of the build was
`title_environment` so the environment was going in empty so the role
name didn't match one that exists in the management account so the
assume role failed because the role doesn't exist.

This is Python so we'll stick with snake case as that's what it was
before.